modified main and arranged several bugs, condfident much is working fine

+++ src/vertex_appearance.h
@@ -42,7 +42,7 @@ namespace std
       betweenness{ std::vector<std::vector<double>>(n, std::vector<double>(T, 0.0)) },
       betweenness_exact{ std::vector<std::vector<double>>(n, std::vector<double>(T, 0.0)) },
       sigma{ std::vector<std::vector<double>>(n, std::vector<double>(T, 0)) },
-      sigmadot{ std::vector<std::vector<int>>(n, std::vector<int>(T, 0)) },
+      sigmadot{ std::vector<std::vector<double>>(n, std::vector<double>(T, 0)) },
       totalSigmaT{ std::vector<std::vector<int>>(n, std::vector<int>(T, 0)) },
       pre{ std::vector<std::vector<std::unordered_set<VertexAppearance>>>(n, std::vector<std::unordered_set<VertexAppearance>>(T)) },
        cur_best{ std::vector<std::vector<double>>(n, std::vector<double>(T, std::numeric_limits<double>::infinity())) },
@@ -62,7 +62,7 @@ namespace std
   std::vector<std::vector<double>> deltasvvt;
   // Numbers of shortest paths from a source to each vertex appearance
   std::vector<std::vector<double>> sigma;
-  std::vector<std::vector<int>> sigmadot;
+  std::vector<std::vector<double>> sigmadot;
   std::vector<std::vector<int>> totalSigmaT;
   // Sets of predecessors for each vertex appearance
   std::vector<std::vector<std::unordered_set<VertexAppearance>>> pre;
